Continued with everything but Board and Main
Made some alterations such as changing Card to be abstract and added
equals methods into it. I really don't like the Card class as it is at
the moment. But it should work for the time being. Did the update
location and calculate distance methods in Player. Currently working on
the Guess class, next steps are commented in.